Ver 0.9 Update some texting

+++ YunTable/include/memstore.h
@@ -21,7 +21,7 @@ public void sort_memstore(Memstore* memstore);
 
 public ResultSet* query_memstore_by_row_key(Memstore* memstore, char* row_key);
 
-public ResultSet* query_memstore_by_timestamp(Memstore* memstore, int begin_timestamp, int end_timestamp);
+public ResultSet* query_memstore_by_timestamp(Memstore* memstore, long long begin_timestamp, long long end_timestamp);
 
 public Memstore* reset_memstore(Memstore *memstore, int flushed_size);
